Move AbstractTestFunctionDescriptorToId from legend-pure
